Suport disassembled ToC linking
The linking works by checking each module for a top-level id that
matches an href in the ToC. Additionally, disassembled files now have
the format of collection_id@version:book_id@version, since the
information was needed for correct linking, and we formatted names
like that on the upload step anyways. i.e. the formatting takes place in
disassembly now, rather than in upload.