Remove use of neb config file in fetch-book
Recent improvements to `neb` allow users to invoke commands without
creating a configuration file. Specifically, this change makes use
of the fact that with those updates, if you pass a fully qualified
hostname to neb as the environmnet and no configuration is available
that matches, it will use the environment value as the cnx hostname.