Update pipelines to call validate-xhtml only once
Instead of validating HTML files throughout the PDF and distribution
pipelines, we'll check the relevant artifacts at the end of each to
reduce the build overheads. The reasons for this are we expect any
errors to be rare, and even if intermediate files have issues we're
not concerned so long as the final outputs pass validation.