Modify link-extras uuid lookup to use redirect header
The initial implementation followed the redirect URL and then used
the url field of the requests Response object to parse the UUID.
However, this can lead to unintentional data being appended to the
UUID (e.g. if there's a fragment in the URI). With this change, we'll
rely instead upon the Location data in the header of the redirect
returned by the server.