Patch math in gdocify for conversion errors
The current version of pandoc and texmath we're using to convert our
XHTML files has conversion errors when <mtext> elements have a
mathvariant attribute as "bold-italic". During pandoc converstion it
generates errors of the following and ends up inserting raw TeX in
the final files:

    [WARNING] Could not convert TeX math '\textit{\textbf{x}}', rendering as TeX:
      \textit{\textbf{x}}
              ^
      unexpected "\\"
      expecting "}", text, "{", "$", "$$", "\\(" or "\\["

The conversion succeeds if the element is <mi> instead, so this
change performs that conversion as a work around.