Add 1e books to canonical list
There are cases where 1e books link to modules in other 1e books
that are not present in 2e versions leading to failures in
link-extras. We're adding 1e books as low priority entries to our
list so the books can be built without changing links to be 2e
equivalent modules.