Support git-based job types in COPS
Job types are not selectable at this time, since there is no pipeline
that supports dequeuing these jobs, but once there are, the
infrastructure and validation is now there to handle them in COPS. The
radio buttons can simply be enabled at that point.

One slight restructure was done in the frontend, and that was to put
data that does not need to be reactive into beforeCreate. Seems to be
the more idiomatic way of handling "constant" data in Vue.