Allow pdf_url to be json specifying multiple links
pdf_url can now be a json array with object elements which each have
'text' and 'href' keys. For backwards compatibility, if an error occurs
parsing pdf_url as json, it is interpreted as being only one element,
where the text is 'View' and the href is the pdf_url string.