Remove check for required fields in bake-book-meta
With some of the fallback logic for the script to work in both
archive and git flows, it's not quite so clear what is required or
not in the binder metadata. Removing the explicit check to avoid
having to maintain this given the ambiguity since it doesn't add much
value anymore.