Update fetch-map-resources for relative image paths
This change modifies the script to use the relative URL in the CNXML
image src to locate used image files instead of assuming all images
to be located in a given input directory. It does, however, leave the
input parameter so we can collect unused resources. This aspect can be
revisited in the future as needed, but doesn't impact functionality in
terms of supported relative image paths.