Fix filtering bug in /jobs/check endpoint
Best explained by example:

x = list(range(1000))

y = [n for n in x[:20] if n % 2 == 0]
z = [n for n in x if n % 2 == 0][:20]

We had y, we wanted z.

Since the /jobs/check endpoint was filtering a subset of all jobs, it
was possible for some job to never appear in the result of /jobs/check

Now the endpoint should filter all jobs and return a subset of the
filter result.