Try to optimize job fetching
Assume that jobs more than a day old will not update. Fetch them from
cache when possible. The cache hash is based on start and end datetimes
used to fetch the jobs. Everyday, the cache should invalidate.

Store the jobs from the last 364 days as json so that we do not need to
perform the slow conversion each time.

Concatenate jobs from today to the end of the list of old jobs. Jobs
from today are always fresh from the database.

Additionally, do not add order_by unless the order_by arg is given.
This has a surprisingly large impact on performance.

Adding response_model to FastAPI routes also has a surprisingly large
impact on performance. Performing the conversion myself made a
measurable difference on its own.