Add git_ref to job schema
Whatever the user entered into the version field when creating a job is
stored in the git_ref field. The new field is used when rerunning jobs.

If git_ref is main, you create a job, modify the repository, and rerun
the job, the second job will use the latest commit on main.

Additionally, the job table has been updated to include
<git_ref>@<commit_sha> in the version field (unless git_ref is the same
as commit_sha, then it just shows commit_sha).

One last change was adding a unique constraint to the commit schema
that was missing. We should only store a commit once per repository.