Job links and worker version (#544)
* That's why it took so long to lint!

Prettier was formatting the bundle too.

* Make caching slightly better for dev env

* Add calculateAge to utils

Finds largest increment of time to express a job's age

* Change job id to string since it is a string

* Add worker version, job age, and links to jobs

Various other small improvements related to style and table width

* Quick fix for test and selector

* Always use hash to show details dialog

This makes it easier for people to get a link to a job and it is DRY

* Fix some selectors I broke

* Add button to job dialog to get a link to the job

* Replace hash in history after processing it

Helps avoid showing the same job again when user reloads the page

* Add placeholder for worker version

* Make ui tests more consistent by adding wait times

This would be possible using playwright's selectors, but it is
unbelievably tedious and error prone.

* Encode xml before parsing with lxml

Unicode strings with encoding declaration are not supported.
Using bytes input fixes the issue.

* Make ui tests fail on unexpected job conclusion

Previously, they only failed when the job unexpected failed. They
should, however, fail on any status that indicates that end of a job if
that status was not expected.