Make timezones required for database datetimes
That means they are required for queries, inserts, etc.

Granted, they were already required for queries because the datetime was stored as utc.

This was only a problem for timestamp values sent from frontend because timestamps do not have timezone.

These timestamps, however, are understood to be utc, so it is safe to assume timezone here.