Only delete previous entries when a new one is added
Before:

When an existing entry included in the to_keep was added again, all other
entries for that book were deleted. Only the entry in to_keep would to remain.

After:

When an existing entry included in the to_keep was added again, update that
entry while leaving other existing entries alone.